split stack
