+++ packages/cdk/lib/construct/index.ts
@@ -11,3 +11,4 @@ export * from './guardrail';
 export * from './speech-to-speech';
 export * from './mcp-api';
 export * from './agent-core';
+export * from './generic-agent-core';
