+++ packages/cdk/lib/generative-ai-use-cases-stack.ts
@@ -20,6 +20,7 @@ import { Agent } from 'generative-ai-use-cases';
 import { UseCaseBuilder } from './construct/use-case-builder';
 import { ProcessedStackInput } from './stack-input';
 import { allowS3AccessWithSourceIpCondition } from './utils/s3-access-policy';
+import { AgentCoreStack } from './agent-core-stack';
 
 export interface GenerativeAiUseCasesStackProps extends StackProps {
   readonly params: ProcessedStackInput;
@@ -28,6 +29,8 @@ export interface GenerativeAiUseCasesStackProps extends StackProps {
   readonly knowledgeBaseDataSourceBucketName?: string;
   // Agent
   readonly agents?: Agent[];
+  // Agent Core
+  readonly agentCoreStack?: AgentCoreStack;
   // Video Generation
   readonly videoBucketRegionMap: Record<string, string>;
   // Guardrail
@@ -134,14 +137,23 @@ export class GenerativeAiUseCasesStack extends Stack {
       mcpEndpoint = mcpApi.endpoint;
     }
 
-    // AgentCore Runtime
-    let agentCore: AgentCore | undefined;
-    if (params.agentCoreEnabled) {
-      agentCore = new AgentCore(this, 'AgentCore', {
-        agentCoreEnabled: params.agentCoreEnabled,
+    // AgentCore Runtime (External runtimes and permissions only)
+    let genericRuntimeArn: string | undefined;
+    let genericRuntimeName: string | undefined;
+
+    // Get generic runtime info from AgentCore stack if it exists
+    if (props.agentCoreStack) {
+      genericRuntimeArn = props.agentCoreStack.deployedGenericRuntimeArn;
+      genericRuntimeName = props.agentCoreStack.getGenericRuntimeConfig()?.name;
+    }
+
+    // Create AgentCore construct for external runtimes and permissions
+    if (params.agentCoreExternalRuntimes.length > 0 || genericRuntimeArn) {
+      new AgentCore(this, 'AgentCore', {
         agentCoreExternalRuntimes: params.agentCoreExternalRuntimes,
         idPool: auth.idPool,
-        fileBucket: api.fileBucket,
+        genericRuntimeArn,
+        genericRuntimeName,
       });
     }
 
@@ -179,11 +191,13 @@ export class GenerativeAiUseCasesStack extends Stack {
       speechToSpeechModelIds: params.speechToSpeechModelIds,
       mcpEnabled: params.mcpEnabled,
       mcpEndpoint,
-      agentCoreEnabled: params.agentCoreEnabled,
-      agentCoreGenericRuntime: agentCore
+      agentCoreEnabled:
+        params.createGenericAgentCoreRuntime ||
+        params.agentCoreExternalRuntimes.length > 0,
+      agentCoreGenericRuntime: genericRuntimeArn
         ? {
-            name: agentCore.getGenericRuntimeConfig().name,
-            arn: agentCore.getDeployedRuntimeArns()[0] || '', // Generic runtime ARN
+            name: genericRuntimeName || 'GenericAgentCoreRuntime',
+            arn: genericRuntimeArn,
           }
         : undefined,
       agentCoreExternalRuntimes: params.agentCoreExternalRuntimes,
@@ -404,14 +418,17 @@ export class GenerativeAiUseCasesStack extends Stack {
     });
 
     new CfnOutput(this, 'AgentCoreEnabled', {
-      value: params.agentCoreEnabled.toString(),
+      value: (
+        params.createGenericAgentCoreRuntime ||
+        params.agentCoreExternalRuntimes.length > 0
+      ).toString(),
     });
 
     new CfnOutput(this, 'AgentCoreGenericRuntime', {
-      value: agentCore
+      value: genericRuntimeArn
         ? JSON.stringify({
-            name: agentCore.getGenericRuntimeConfig().name,
-            arn: agentCore.getDeployedRuntimeArns()[0] || '',
+            name: genericRuntimeName || 'GenericAgentCoreRuntime',
+            arn: genericRuntimeArn,
           })
         : 'null',
     });
