+++ packages/cdk/lib/stack-input.ts
@@ -137,7 +137,8 @@ const baseStackInputSchema = z.object({
     .default([]),
   inlineAgents: z.boolean().default(false),
   // Agent Core Runtime
-  agentCoreEnabled: z.boolean().default(false),
+  createGenericAgentCoreRuntime: z.boolean().default(false),
+  agentCoreRegion: z.string().nullish(),
   agentCoreExternalRuntimes: z
     .array(
       z.object({
